[RM] Bug #58956. Fix database mysql shell and bat scripts

+++ FrontEnd/SailfishFrontEnd/etc/DB/create_mysql_db.sh
@@ -53,7 +53,15 @@ checkDatabaseAvailability()
 
 createDatabase()
 {
-    mysql --host=${HOST} --port=${PORT} --user=${SUPERUSER} --password=${SUPERPASSWORD} -e 'SELECT VERSION();' | grep '| 5'
+    output=$(mysql --host=${HOST} --port=${PORT} --user=${SUPERUSER} --password=${SUPERPASSWORD} -e 'SELECT VERSION();')
+    local IS_LOGIN=$?
+    if [ "$IS_LOGIN" -ne "0" ]; then
+        echo "Error determining mysql version"
+        echo $output
+        return 1
+    fi
+
+    echo "$output" | grep '^5.'
     local IS_FIFTH_VERSION=$?
 
     QUERY="drop database if exists $DATABASE;
@@ -72,7 +80,7 @@ createDatabase()
 
             grant all on $DATABASE.* to '$USER'@'localhost';
             grant all on $DATABASE.* to '$USER'@'localhost.localdomain';"
-    fi
+     fi
 
     mysql --host=${HOST} --port=${PORT} --user=${SUPERUSER} --password=${SUPERPASSWORD} -e "$QUERY"
 }
